Filter: Fixed some tests.
Chrome plays all transitions when the page is rendered. If the filter
is in dialog mode a user sees the closing filter dialog animation
first. *argh* This leads to unstable tests because sometimes Protractor
will complain that it clicked the open filter button which was still
covered by the shrinking dialog…